ModularForms.jl Documentation

ModularForms.jl Documentation

ModularForms.jl Documentation

Functions

Index

Functions

ModularForms.delta_qexp — Function.

delta_qexp(prec=10, var="q", K=ZZ)

Return the q-expansion of the normalized cusp form of weight 12 to precision prec as a power series over K in the variable var.

Arguments

prec::Integer=10: precision of the output

var::String="q": variable name

K=ZZ: base ring of the output

Examples still missing

source

ModularForms.delta_k_qexp — Function.

delta_k_qexp(k, prec=10, var="q")

Return the q-expansion of the unique normalized eigenform of weight k and level 1 for k in {12, 16, 18, 20, 22, 26} as a power series to precision prec in var over ZZ.

These eigenforms are the normalized generators for the six one-dimensional spaces of cusp forms of level 1.

Arguments

k::Integer: weight of the unique normalized eigenform

prec::Integer=10: precision of the output

var::String="q": variable name

Examples still missing

source

ModularForms.eisenstein_series_qexp — Function.

eisenstein_series_qexp(k, prec=10, K=QQ, var="q", normalization="linear")

Return the q-expansion of the normalized weight k Eisenstein series on the modular group to precision prec as a power series in the ring K in variable var, using the given normalization.

Three normalizations are available: "linear" (default), "constant", and "integral". If the normalization is "linear" then the linear coefficient is 1. If it is "constant" then the series will be normalized to have constant term 1. If the normalization is "integral" then the series will be normalized to have integer coefficients and no common factors. Note: To prevent errors, the output will be in the ring QQ if the normalization is "linear" or "constant".

Arguments

k::Integer: even positive integer, weight of the Eisenstein series

prec::Integer=10: precision of the output

K=ZZ: base ring of the output

var::String="q": variable name

normalization::String="linear": normalization to use

Examples still missing

source

ModularForms.eta_quotient — Function.

eta_quotient(g, prec=10)

Return the eta-quotient relative to the given collection of integers g as a power series up to precision prec.

The input g is an array of pairs [[t1,r1], [t2,r2], ..., [ts,rs]] where each tj is a positive integer and each rj a nonnegative integer. An error is thrown if the sum of tj*rj for j in {1, ..., s} does not equal 24, as the function only applies to eta-quotients that are cusp forms. The output is a cusp form of integral weight.

Arguments

g::Array{Array}: an array of pairs, collection of integers

prec::Integer=10: precision of the output

Examples still missing

source

ModularForms.victor_miller_basis — Function.

victor_miller_basis_poly(k, prec=10, cusp_only=false, var="q")

Return the Victor Miller basis for modular forms of weight k and level 1 to precision prec as an array whose entries are power series in ZZ[[var]]. If cusp_only is true, then return only a basis for the cuspidal subspace.

The algorithm uses the proof of Lemma 2.20 from William A. Stein.

Arguments

k::Integer: weight

prec::Integer=10: precision of the output

cusp_only::Boolean=false

var::String="q": variable name

Examples still missing

source

ModularForms.hecke_operator_on_qexp — Function.

hecke_operator_on_qexp(f, n, k, prec=nothing)

Compute the image of the q-expansion f of a modular form under the Hecke operator T_n of weight k. Return a power series to precision prec.

Arguments

f::RelSeriesElem: q-expansion

n::Integer: integer >=1

k::Integer: weight

prec::Integer=nothing: precision of the output

Examples still missing

source

ModularForms.hecke_operator_on_basis — Function.

hecke_operator_on_basis(B, n, k)

Compute the matrix of the Hecke operator T_n of weight k relative to the given basis B of q-expansions for a space of modular forms.

Arguments

B::Array: array of q-expansions

n::Integer: integer >=1

k::Integer: weight

Examples still missing

source

ModularForms.prime_range — Function.

prime_range(n)

Return an array consisting of all primes up to and including n.

source

ModularForms.poly_to_power_series — Function.

poly_to_power_series(f, K, prec=10)

Convert the polynomial f over K to a relative power series to precision prec as a polynomial.

The power series has the same coefficient ring and variable name as f. The input polynomial f must be correct up to precision prec.

Arguments

' f::PolyElem: polynomial over K, correct up to prec

K: base ring of the output

prec::Integer=10: precision of the output

Examples still missing

source

Index

ModularForms.delta_k_qexp

ModularForms.delta_qexp

ModularForms.eisenstein_series_qexp

ModularForms.eta_quotient

ModularForms.hecke_operator_on_basis

ModularForms.hecke_operator_on_qexp

ModularForms.poly_to_power_series

ModularForms.prime_range

ModularForms.victor_miller_basis